VideoStreamingBackend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Database
sqlalchemy>=2.0.0
pg8000>=1.30.0
alembic>=1.13.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Security / JWT
python-jose[cryptography]>=3.3.0

# Video processing
opencv-python-headless>=4.9.0
numpy>=1.26.0

# Stream URL resolution (YouTube, HLS, etc.)
yt-dlp>=2024.1.0
streamlink>=6.0.0

# Async utilities
asyncio-throttle>=1.0.2
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
